backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.31.0
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
reportlab>=4.4.10
pytest==8.3.3
pytest-cov==5.0.0
